--hfst to expand trimmed binary, skipping tagger
New optional switch --hfst will use hfst-fst2strings, and skip any
tagger/CG (so assumes no substitute/mapping rules affecting testvoc);
runs about 10x faster. E.g.
./dev/testvoc/generation.sh --hfst dan-nor
or
./dev/testvoc/generation.sh --hfst dan-nor dan-nor-alternative.automorf.bin

+++ dev/testvoc/generation.sh
@@ -2,6 +2,13 @@
 
 set -e -u
 
+if [[ $# -ge 1 && $1 = --hfst ]]; then
+    HFST=true
+    shift
+else
+    HFST=false
+fi
+
 if [[ $# -eq 1 ]]; then
     mode=$1
     dix=guess
@@ -12,15 +19,25 @@ else
     cat >&2 <<EOF
 Usage: $0 lang1-lang2
 or:    $0 lang1-lang2 foo.dix
+or:    $0 --hfst lang1-lang2 foo-bar.automorf.bin
 
 Replaces the first step of the pipeline with the expanded analyser and
 shows the resulting generation errors.
 
-For example, do \`$0 nno-nob' in trunk/apertium-nno-nob/'
-to find generation errors in the nno-nob direction (assumes that
-modes/nno-nob.mode exists). If the source .dix file has a non-standard
-name, you can specify it in the second argument, for example
-\`$0 eng-sco ../apertium-eng_feil/apertium-eng.eng.dix'
+For example, do \`$0 nno-nob' in trunk/apertium-nno-nob/' to find
+generation errors in the nno-nob direction (assumes that
+modes/nno-nob.mode exists).
+
+If the source .dix file has a non-standard name, you can specify it in
+the second argument, for example \`$0 eng-sco
+../apertium-eng_feil/apertium-eng.eng.dix'
+
+If you pass --hfst, the trimmed analyser will be used, and
+disambiguation will be skipped. This is a lot faster, but assumes you
+don't use mapping rules in CG. With --hfst, the optional dix argument
+is interpreted as the analyser binary (defaulting to
+lang1-lang2.automorf.bin).
+
 EOF
     exit 1
 fi
@@ -37,6 +54,25 @@ analysis_expansion () {
             gsub("$","\\$",esc)
             print "["esc"] ^"$1"/"$2"$ ^./.<sent>"clb"$"
           }'
+}
+
+analysis_expansion_hfst () {
+    lt-print "$1" \
+        | sed 's/ /@_SPACE_@/g' \
+        | hfst-txt2fst -e ε \
+        | hfst-project -p lower \
+        | hfst-fst2strings  -c0  \
+        | awk -v clb="$2" '
+          /[][$^{}\\]/{next} # skip escaping hell
+          {
+            gsub("]","\\]")
+            esc=$0
+            gsub("/","\\/",esc)
+            gsub("^","\\^",esc)
+            gsub("$","\\$",esc)
+            print "["esc"] ^"$0"$ ^.<sent>"clb"$"
+          }'
+    # give the "disambiguated" output, no forms
 }
 
 split_ambig () {
@@ -63,6 +99,17 @@ mode_after_analysis ()
                   sed 's/\$1/-d/g;s/\$2//g')
 }
 
+mode_after_tagger ()
+{
+    eval $(grep '|' "$1" |\
+                  sed 's/[^|]*|//' |\
+                  sed 's/.*apertium-pretransfer/apertium-pretransfer/' |\
+                  sed 's/lt-proc -p[^|]*/cat/' |\
+                  sed 's/autobil.bin *|/& split_ambig |/' |\
+                  sed 's/\$1/-d/g;s/\$2//g')
+    # lt-proc -p fails, that's why we remove that
+}
+
 only_errs () {
     grep '][^<]*[#/]'
 }
@@ -70,21 +117,28 @@ only_errs () {
 
 lang1=${mode%%-*}
 
-if [[ ${dix} = guess ]]; then
-    lang1dir=$(grep -m1 "^AP_SRC.*apertium-${lang1}" config.log | sed "s/^[^=]*='//;s/'$//")
-    dix=${lang1dir}/apertium-${lang1}.${lang1}.dix
-fi
-
 clb=""
 case ${lang1} in
     nno|nob) clb="<clb>" ;;
 esac
 
-# Make it possible to edit the .dix while testvoc is running:
-dixtmp=$(mktemp -t gentestvoc.XXXXXXXXXXX)
-trap 'rm -f "${dixtmp}"' EXIT
-cat "${dix}" > "${dixtmp}"
-
-analysis_expansion "${dixtmp}" "${clb}" \
-    | mode_after_analysis modes/"${mode}".mode \
-    | only_errs
+if $HFST; then
+    if [[ ${dix} = guess ]]; then
+        dix="${mode}".automorf.bin
+    fi
+    analysis_expansion_hfst "${dix}" "${clb}" \
+        | mode_after_tagger modes/"${mode}".mode \
+        | only_errs
+else
+    if [[ ${dix} = guess ]]; then
+        lang1dir=$(grep -m1 "^AP_SRC.*apertium-${lang1}" config.log | sed "s/^[^=]*='//;s/'$//")
+        dix=${lang1dir}/apertium-${lang1}.${lang1}.dix
+    fi
+    # Make it possible to edit the .dix while testvoc is running:
+    dixtmp=$(mktemp -t gentestvoc.XXXXXXXXXXX)
+    trap 'rm -f "${dixtmp}"' EXIT
+    cat "${dix}" > "${dixtmp}"
+    analysis_expansion "${dixtmp}" "${clb}" \
+        | mode_after_analysis modes/"${mode}".mode \
+        | only_errs
+fi
